+++ docs/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+* Resonant Element search filters such as `deepsight:ruinous` have been removed as these currencies
+are now deprecated.
 * Selected Super ability is now displayed on Solar subclass icons.
 
 ## 7.18.1 <span class="changelog-date">(2022-05-24)</span>